+++ Application/Core/Events.php
@@ -44,119 +44,6 @@ class Events
             }
         }
 
-        self::addContentSnippetOptIn();
-        self::addContentSnippetOptOut();
-        self::addContentSnippetUpdate();
-    }
-
-    /**
-     * Add content snippet for Opt-In case
-     */
-    protected static function addContentSnippetOptIn()
-    {
-        $text = <<<'EOT'
-Die [FIRMA] erstellt auf dieser Webseite pseudonyme Nutzerprofile auf Basis Ihres Online-Nutzungsverhaltens und nutzt dazu Cookies.
-Dieses Profil ermöglicht es uns, Sie online so ausführlich zu beraten, wie es ein Verkäufer in einem persönlichen Gespräch in einem [FIRMA] Store kann.
-Weiterhin können wir unser Angebot und unsere Werbung auf Partnerseiten an Ihre persönlichen Bedürfnisse anpassen,
-sodass für Sie relevante Produkte angezeigt und uninteressante Angebote ausgeblendet werden.
-Die Verarbeitung erfolgt gemäß Art. 6 Abs. 1 lit. f DSGVO und Sie können jederzeit von Ihren Betroffenenrechten Gebrauch machen.
-Wenn Sie dies nicht möchten, können Sie hier widersprechen. Wenn Sie unsicher sind, finden Sie hier die gesamten Datenschutzhinweise.
-<div>
-    <a class="oepersonalization-optin" href="#" data-dismiss="alert">Ich bin einverstanden (Personalisierung aktivieren)</a>
-</div>
-EOT;
-        $sql = "select count(oxid) from `oxcontents` where oxloadid = 'oepersonalizationoptin'";
-        $result = DatabaseProvider::getDb()->getCol($sql);
-        if ($result[0] == 0) {
-            $id = Registry::getUtilsObject()->generateUId();
-            $content = oxNew(Content::class);
-            $content->setId($id);
-            $content->setLanguage(0);
-            $content->oxcontents__oxloadid = new Field('oepersonalizationoptin');
-            $content->oxcontents__oxtitle = new Field('Cookie "Ich bin einverstanden (Personalisierung aktivieren)" Hinweis');
-            $content->oxcontents__oxcontent = new Field($text);
-            $content->save();
-
-            $content->setLanguage(1);
-            $content->oxcontents__oxtitle = new Field('Cookie "I agree (activate personalization)" hint');
-            $content->oxcontents__oxcontent = new Field('Please update this text. An example can be found in German language of this entry.');
-            $content->save();
-        }
-    }
-
-    /**
-     * Add content snippet for Opt-Out case
-     */
-    protected static function addContentSnippetOptOut()
-    {
-        $text = <<<'EOT'
-Die [FIRMA] erstellt auf dieser Webseite pseudonyme Nutzerprofile auf Basis Ihres Online-Nutzungsverhaltens und nutzt dazu Cookies.
-Dieses Profil ermöglicht es uns, Sie online so ausführlich zu beraten, wie es ein Verkäufer in einem persönlichen Gespräch in einem [FIRMA] Store kann.
-Weiterhin können wir unser Angebot und unsere Werbung auf Partnerseiten an Ihre persönlichen Bedürfnisse anpassen,
-sodass für Sie relevante Produkte angezeigt und uninteressante Angebote ausgeblendet werden.
-Die Verarbeitung erfolgt gemäß Art. 6 Abs. 1 lit. f DSGVO und Sie können jederzeit von Ihren Betroffenenrechten Gebrauch machen.
-Wenn Sie dies nicht möchten, können Sie hier widersprechen. Wenn Sie unsicher sind, finden Sie hier die gesamten Datenschutzhinweise.
-<div>
-    <a class="oepersonalization-optout" href="#" data-dismiss="alert">Widerspruch (Personalisierung deaktivieren)</a>
-</div>
-EOT;
-        $sql = "select count(oxid) from `oxcontents` where oxloadid = 'oepersonalizationoptout'";
-        $result = DatabaseProvider::getDb()->getCol($sql);
-        if ($result[0] == 0) {
-            $id = Registry::getUtilsObject()->generateUId();
-            $content = oxNew(Content::class);
-            $content->setId($id);
-            $content->setLanguage(0);
-            $content->oxcontents__oxloadid = new Field('oepersonalizationoptout');
-            $content->oxcontents__oxtitle = new Field('Cookie "Widerspruch (Personalisierung deaktivieren)" Hinweis');
-            $content->oxcontents__oxcontent = new Field($text);
-            $content->save();
-
-            $content->setLanguage(1);
-            $content->oxcontents__oxtitle = new Field('Cookie "I disagree (deactivate personalization)" hint');
-            $content->oxcontents__oxcontent = new Field('Please update this text. An example can be found in German language of this entry.');
-            $content->save();
-        }
-    }
-
-    /**
-     * Add content snippet to update privacy settings
-     */
-    protected static function addContentSnippetUpdate()
-    {
-        $textGerman = <<<'EOT'
-<div id="oepersonalization-update">
-    <h4>Tracking</h4>
-    <input type="radio" name="oepersonalization-state" value="ALLOW"> Zulassen
-    <input type="radio" name="oepersonalization-state" value="DENY"> Verbieten
-    <div><button type="button" class="btn btn-primary">Aktualisieren</button></div>
-</div>
-EOT;
-        $textEnglish = <<<'EOT'
-<div id="oepersonalization-update">
-    <h4>Tracking</h4>
-    <input type="radio" name="oepersonalization-state" value="ALLOW"> Allow
-    <input type="radio" name="oepersonalization-state" value="DENY"> Deny
-    <div><button type="button" class="btn btn-primary">Update</button></div>
-</div>
-EOT;
-        $sql = "select count(oxid) from `oxcontents` where oxloadid = 'oepersonalizationupdate'";
-        $result = DatabaseProvider::getDb()->getCol($sql);
-        if ($result[0] == 0) {
-            $id = Registry::getUtilsObject()->generateUId();
-            $content = oxNew(Content::class);
-            $content->setId($id);
-            $content->setLanguage(0);
-            $content->oxcontents__oxloadid = new Field('oepersonalizationupdate');
-            $content->oxcontents__oxtitle = new Field('Privacy Protection-Einstellungen');
-            $content->oxcontents__oxcontent = new Field($textGerman);
-            $content->save();
-
-            $content->setLanguage(1);
-            $content->oxcontents__oxtitle = new Field('Privacy protection settings');
-            $content->oxcontents__oxcontent = new Field($textEnglish);
-            $content->save();
-        }
     }
 
     /**
@@ -258,16 +145,22 @@ EOT;
                 'value' => ''
             ],
             [
-                'group' => '',
-                'name' => 'sOePersonalizationTrackingShowNote',
-                'type' => 'str',
-                'value' => 'no'
-            ],
-            [
                 'group' => 'oepersonalization_export',
                 'name' => 'sOePersonalizationExportPath',
                 'type' => 'str',
                 'value' => 'export/oepersonalization'
+            ],
+            [
+                'group' => '',
+                'name' => 'oePersonalizationActive',
+                'type' => 'bool',
+                'value' => ''
+            ],
+            [
+                'group' => '',
+                'name' => 'oePersonalizationContainerId',
+                'type' => 'str',
+                'value' => ''
             ],
         ];
     }