+++ Application/views/admin/de/module_options.php
@@ -31,13 +31,10 @@ $aLang = [
     'SHOP_MODULE_sOePersonalizationWidgetTemplateThankYouPage' => 'Widget-Template',
 
     'SHOP_MODULE_blOePersonalizationEnableTracking' => 'econda Analytics aktivieren',
+    'SHOP_MODULE_oePersonalizationActive' => 'Personalisierung aktivieren',
+    'SHOP_MODULE_oePersonalizationContainerId' => 'ID des econda ARP Containers',
 
     'SHOP_MODULE_blOePersonalizationEnableTagManager' => 'Tag Manager aktivieren',
-
-    'SHOP_MODULE_sOePersonalizationTrackingShowNote' => 'Hinweis',
-    'SHOP_MODULE_sOePersonalizationTrackingShowNoteNo' => 'Nicht anzeigen',
-    'SHOP_MODULE_sOePersonalizationTrackingShowNoteOptIn' => '"Ich bin einverstanden (Personalisierung aktivieren)" anzeigen',
-    'SHOP_MODULE_sOePersonalizationTrackingShowNoteOptOut' => '"Widerspruch (Personalisierung deaktivieren)" anzeigen',
 
     'SHOP_MODULE_GROUP_oepersonalization_export' => 'Export',
     'SHOP_MODULE_sOePersonalizationExportPath'=> 'Export-Pfad',