+++ Tests/Codeception/_data/oepersonalizationData.php
@@ -1,0 +1,14 @@
+<?php
+
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+/**
+ * Returns users data for Personalization testing
+ */
+return array(
+    'oePersonalizationAccountId' => '<oePersonalization Account Id>',
+    'econdaARPContainerId' => '<econda ARP Container Id>'
+);